requirements.txt
openai
python-dotenv
colorama
SpeechRecognition
requests
youtube-search-python
pyautogui
pyttsx3
redis
